+++ modules/mrc_media/src/Plugin/EntityBrowser/Widget/DropzoneUpload.php
@@ -2,17 +2,12 @@
 
 namespace Drupal\mrc_media\Plugin\EntityBrowser\Widget;
 
-use Drupal\Component\Utility\Bytes;
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\Core\Render\Element;
 use Drupal\Core\Session\AccountProxyInterface;
 use Drupal\dropzonejs\DropzoneJsUploadSave;
-use Drupal\entity_browser\WidgetBase;
 use Drupal\entity_browser\WidgetValidationManager;
-use Drupal\Component\Utility\NestedArray;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\file\Entity\File;
-use Drupal\inline_entity_form\ElementSubmit;
 use Drupal\media\Entity\Media;
 use Drupal\mrc_media\BundleSuggestion;
 use Symfony\Component\DependencyInjection\ContainerInterface;
@@ -60,7 +55,6 @@ class DropzoneUpload extends MediaBrowserBase {
       $container->get('current_user')
     );
   }
-
 
   /**
    * {@inheritdoc}
@@ -184,31 +178,6 @@ class DropzoneUpload extends MediaBrowserBase {
         $media_entity->save();
       }
     }
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function submit(array &$element, array &$form, FormStateInterface $form_state) {
-    parent::submit($element, $form, $form_state);
-
-    $children = Element::children($element['entities']);
-    foreach ($children as $child) {
-      $entity_form = $element['entities'][$child];
-
-      if (!isset($entity_form['#ief_element_submit'])) {
-        continue;
-      }
-
-      foreach ($entity_form['#ief_element_submit'] as $submit_function) {
-        call_user_func_array($submit_function, [&$entity_form, $form_state]);
-      }
-    }
-
-    $media_entities = $this->prepareEntities($form, $form_state);
-
-    $this->selectEntities($media_entities, $form_state);
-    $form_state->set(['dropzonejs', $this->uuid(), 'files'], []);
   }
 
   /**
